+++ neurons/miners/zephyr/miner.py
@@ -23,8 +23,8 @@ import bittensor as bt
 # Bittensor Miner Template:
 import prompting
 from prompting.protocol import PromptingSynapse
-from prompting.llms.hf_llm import load_hf_pipeline, HuggingFaceLLM
-from prompting.llms.hf_llm import HuggingFaceLLM
+from prompting.llms.hf import load_hf_pipeline, HuggingFaceLLM
+from prompting.llms.hf import HuggingFaceLLM
 
 # import base miner class which takes care of most of the boilerplate
 from neurons.miner import Miner
