+++ prompting/llms/__init__.py
@@ -1,3 +1,3 @@
 from .base_llm import BasePipeline, BaseLLM
-from .hf_llm import HuggingFacePipeline, HuggingFaceLLM
+from .hf import HuggingFacePipeline, HuggingFaceLLM
 from .vllm_llm import vLLM_LLM, vLLMPipeline, load_vllm_pipeline
